Forms: Add form response webhooks (#46059)
* Add form webhook settings panel

Added under a feature flag to test the webhook implementation.

* Add a form webhook class

This is an exploratory commit where I'm using a different class, based
on the Post_To_Url one, to process and dispatch the webhooks on form
submission.

* changelog

* Add webhook id input field

I added the webhook id input to the settings section to explore how we
could use it for advanced users that would like to programmatically
manipulate form data before sending it to webhook. For that, we'll need
to expose a filter which I'll bring in a follow up commit.

* Add jetpack_forms_before_webhook_request

I'm adding this filter here to allow developers to intercept and modify
the form data that I'm sending in the webhook request. This should give
enough fleixibility to advanced users.

* Log response and errors to post meta table

* Fix param type in documentation

* Fix JetpackContactFormAttributes type declaration

* Fix unit test

* Forms: add comprehensive webhook functionality (#46099)

* let salesforce process remain as it was. Split (just in case) postToUrl and migrate it to a webhook

* add class consts for cleaner code

* create and validate webhook setup with class consts, ditch non matching formats/methods

* use class constants on request process

* invert filter params order so data is returned by default even when no filter is set

* add tests for Form_Webhooks class

* make phan happy on test files

* changelog

* Fix webhook method validation

* Forms: Add webhook logging and initialization flag (#46110)

* add logging for skipped webhooks

* only init webhooks class if webhooks are enabled

* changelog

* add tests for webhooks and contact form flags

* fix phan issues

* Add plan-based feature support for form webhooks (#46138)

* add form-webhooks support on free JP plans

* set flag form-webhooks based on current plan support

* consider both filter and feature flag for enabling webhook settings

* add webhooks feature support plans

* add changelogs

* add wpcomsh changelog

---------

Co-authored-by: Christian Gastrell <cgastrell@gmail.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/19829614670

Upstream-Ref: Automattic/jetpack@92ec44c150a00bc270fe7b675630c9c653a25d23